add state retry and catcher using ginkgo and gomega to test
optimize code

+++ core/error.go
@@ -3,7 +3,7 @@ package core
 type (
 	Retry struct {
 		ErrorEquals     []string
-		IntervalSeconds int32
+		IntervalSeconds float32
 		MaxAttempts     int32
 		// A number that is the multiplier by which the retry interval increases on each attempt (default 2.0).
 		BackoffRate float32
